+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './packageReader';
-export * from './npmClient';
+export * from './ncuClient';
 export * from './packageManager';
 export * from './versionUtils';
 export * from './notify';